visioncop/tests/test_api.py
import pytest
from fastapi.testclient import TestClient
from app.main import app
import os
from PIL import Image
import numpy as np

client = TestClient(app)

# Create a dummy image for testing
def create_dummy_image():
    """Create a test image file."""
    img = Image.new('RGB', (100, 100), color=(73, 109, 137))
    img.save("test_image.jpg")
    return "test_image.jpg"

def cleanup_dummy_image(filename):
    """Remove test files."""
    if os.path.exists(filename):
        os.remove(filename)

def test_root_get():
    """Test root endpoint (if exists)."""
    # Assuming no root endpoint, this will fail as expected
    response = client.get("/")
    assert response.status_code == 404  # Not Found

def test_index_endpoint():
    """Test image indexing endpoint."""
    test_file = create_dummy_image()
    
    try:
        with open(test_file, "rb") as f:
            response = client.post("/index",
                                 files={"file": ("test_image.jpg", f, "image/jpeg")})
        
        assert response.status_code == 200
        data = response.json()
        assert "message" in data
        assert "indexed successfully" in data["message"]
    finally:
        cleanup_dummy_image(test_file)
        # Also cleanup from data directory if created
        if os.path.exists("data/images/test_image.jpg"):
            os.remove("data/images/test_image.jpg")

def test_search_endpoint():
    """Test image search endpoint."""
    test_file = create_dummy_image()
    
    try:
        with open(test_file, "rb") as f:
            response = client.post("/search",
                                 files={"file": ("test_image.jpg", f, "image/jpeg")})
        
        assert response.status_code == 200
        data = response.json()
        assert "similar_images" in data
        assert isinstance(data["similar_images"], list)
    finally:
        cleanup_dummy_image(test_file)

# Placeholder for more tests
def test_image_verification():
    """Stub: Test image verification functions."""
    pass

def test_database_operations():
    """Stub: Test database CRUD operations."""
    pass
